chore: add size action
