+++ .github/workflows/lerna-ci.yml
@@ -1,4 +1,4 @@
-name: Lerna CI
+name: Continuous Integration
 
 on:
   pull_request:
@@ -19,9 +19,9 @@ jobs:
         with:
           node-version: 14
           check-latest: true
-          registry-url: 'https://registry.npmjs.org/'
-          scope: '@talend'
-          cache: 'yarn'
+          registry-url: "https://registry.npmjs.org/"
+          scope: "@talend"
+          cache: "yarn"
 
       - name: Install
         run: yarn install --frozen-lockfile
@@ -29,13 +29,18 @@ jobs:
       - name: Test
         run: yarn test
 
-      - uses: preactjs/compressed-size-action@v2
-        if: github.ref != 'refs/heads/master'
-        with:
-          repo-token: '${{ secrets.GITHUB_TOKEN }}'
-          build-script: 'pre-release'
-          pattern: './packages/*/dist/*.{js,css,json}'
-          exclude: '{**/*.html,**/*.map,**/node_modules/**}'
-          compression: 'none'
-          minimum-change-threshold: 100
+      - name: size check
+        id: sizecheck
+        run: |
+          node size.js
+          echo "::set-output name=diff::$(git status --short size.json)"
+        env:
+          THRESHOLD: 100;
 
+      - name: Commit size change
+        if: steps.sizecheck.outputs.diff
+        uses: actions-js/push@master
+        with:
+          github_token: ${{ secrets.GITHUB_TOKEN }}
+          message: "chore: update size"
+          branch: ${{ github.head_ref }}
