+++ size.json
@@ -1,0 +1,87 @@
+{
+  "packages/cmf-cqrs/dist/TalendReactCmfCqrs.js": 35615,
+  "packages/cmf-cqrs/dist/TalendReactCmfCqrs.js.dependencies.json": 539,
+  "packages/cmf-cqrs/dist/TalendReactCmfCqrs.min.js": 6810,
+  "packages/cmf-cqrs/dist/TalendReactCmfCqrs.min.js.dependencies.json": 558,
+  "packages/cmf-router/dist/TalendReactCmfRouter.js": 179935,
+  "packages/cmf-router/dist/TalendReactCmfRouter.js.dependencies.json": 834,
+  "packages/cmf-router/dist/TalendReactCmfRouter.min.js": 26818,
+  "packages/cmf-router/dist/TalendReactCmfRouter.min.js.dependencies.json": 864,
+  "packages/cmf/dist/TalendReactCmf.js": 535366,
+  "packages/cmf/dist/TalendReactCmf.js.dependencies.json": 1253,
+  "packages/cmf/dist/TalendReactCmf.min.js": 110001,
+  "packages/cmf/dist/TalendReactCmf.min.js.dependencies.json": 1294,
+  "packages/components/dist/TalendReactComponents.css": 147641,
+  "packages/components/dist/TalendReactComponents.js": 3151428,
+  "packages/components/dist/TalendReactComponents.js.dependencies.json": 2855,
+  "packages/components/dist/TalendReactComponents.min.js": 701590,
+  "packages/components/dist/TalendReactComponents.min.js.dependencies.json": 2932,
+  "packages/containers/dist/TalendReactContainers.css": 1500,
+  "packages/containers/dist/TalendReactContainers.js": 442295,
+  "packages/containers/dist/TalendReactContainers.js.dependencies.json": 1411,
+  "packages/containers/dist/TalendReactContainers.min.js": 95183,
+  "packages/containers/dist/TalendReactContainers.min.js.dependencies.json": 1454,
+  "packages/datagrid/dist/TalendReactDatagrid.css": 65380,
+  "packages/datagrid/dist/TalendReactDatagrid.js": 302959,
+  "packages/datagrid/dist/TalendReactDatagrid.js.dependencies.json": 1226,
+  "packages/datagrid/dist/TalendReactDatagrid.min.js": 60855,
+  "packages/datagrid/dist/TalendReactDatagrid.min.js.dependencies.json": 1263,
+  "packages/dataviz/dist/TalendReactDataviz.css": 4597,
+  "packages/dataviz/dist/TalendReactDataviz.js": 818502,
+  "packages/dataviz/dist/TalendReactDataviz.js.dependencies.json": 853,
+  "packages/dataviz/dist/TalendReactDataviz.min.js": 598519,
+  "packages/dataviz/dist/TalendReactDataviz.min.js.dependencies.json": 876,
+  "packages/design-system/dist/TalendDesignSystem.css": 1121623,
+  "packages/design-system/dist/TalendDesignSystem.js": 1120893,
+  "packages/design-system/dist/TalendDesignSystem.js.dependencies.json": 1176,
+  "packages/design-system/dist/TalendDesignSystem.min.js": 446622,
+  "packages/design-system/dist/TalendDesignSystem.min.js.dependencies.json": 1205,
+  "packages/design-tokens/dist/TalendDesignTokens.css": 29488,
+  "packages/design-tokens/dist/TalendDesignTokens.js": 29590,
+  "packages/design-tokens/dist/TalendDesignTokens.js.dependencies.json": 2,
+  "packages/design-tokens/dist/TalendDesignTokens.min.js": 24504,
+  "packages/design-tokens/dist/TalendDesignTokens.min.js.dependencies.json": 2,
+  "packages/faceted-search/dist/TalendReactFacetedSearch.css": 10508,
+  "packages/faceted-search/dist/TalendReactFacetedSearch.js": 368621,
+  "packages/faceted-search/dist/TalendReactFacetedSearch.js.dependencies.json": 995,
+  "packages/faceted-search/dist/TalendReactFacetedSearch.min.js": 92300,
+  "packages/faceted-search/dist/TalendReactFacetedSearch.min.js.dependencies.json": 1025,
+  "packages/flow-designer/dist/ReactFlowDesigner.min.js": 91151,
+  "packages/flow-designer/dist/ReactFlowDesigner.min.js.dependencies.json": 854,
+  "packages/forms/dist/TalendReactForms.css": 16974,
+  "packages/forms/dist/TalendReactForms.js": 1350166,
+  "packages/forms/dist/TalendReactForms.js.dependencies.json": 1773,
+  "packages/forms/dist/TalendReactForms.min.js": 418850,
+  "packages/forms/dist/TalendReactForms.min.js.dependencies.json": 1821,
+  "packages/http/dist/TalendHttp.js": 24792,
+  "packages/http/dist/TalendHttp.js.dependencies.json": 2,
+  "packages/http/dist/TalendHttp.min.js": 4719,
+  "packages/http/dist/TalendHttp.min.js.dependencies.json": 2,
+  "packages/icons/dist/TalendIcons.js": 624509,
+  "packages/icons/dist/TalendIcons.js.dependencies.json": 128,
+  "packages/icons/dist/TalendIcons.min.js": 552674,
+  "packages/icons/dist/TalendIcons.min.js.dependencies.json": 131,
+  "packages/icons/dist/bundle.js": 6313,
+  "packages/icons/dist/info.js": 10714,
+  "packages/icons/dist/react.esm.js": 529095,
+  "packages/icons/dist/react.js": 568530,
+  "packages/icons/dist/talend-icons-webfont.css": 16627,
+  "packages/icons/dist/talendicons.css": 334,
+  "packages/jsfc/dist/index.js": 95516,
+  "packages/jsfc/dist/index.js.dependencies.json": 2,
+  "packages/router-bridge/dist/TalendRouterBridge.js": 152444,
+  "packages/router-bridge/dist/TalendRouterBridge.js.dependencies.json": 770,
+  "packages/router-bridge/dist/TalendRouterBridge.min.js": 35433,
+  "packages/router-bridge/dist/TalendRouterBridge.min.js.dependencies.json": 790,
+  "packages/sagas/dist/TalendReactSagas.js": 12113,
+  "packages/sagas/dist/TalendReactSagas.js.dependencies.json": 349,
+  "packages/sagas/dist/TalendReactSagas.min.js": 2134,
+  "packages/sagas/dist/TalendReactSagas.min.js.dependencies.json": 361,
+  "packages/stepper/dist/TalendReactStepper.js": 26317,
+  "packages/stepper/dist/TalendReactStepper.js.dependencies.json": 833,
+  "packages/stepper/dist/TalendReactStepper.min.js": 4856,
+  "packages/stepper/dist/TalendReactStepper.min.js.dependencies.json": 855,
+  "packages/theme/dist/bootstrap.css": 149130,
+  "packages/theme/dist/bootstrap.js": 4327,
+  "packages/theme/dist/bootstrap.js.dependencies.json": 3
+}
